create new EventListener API for receiving events
The existing EventHandler API has some severe limitations:

* there can only be a single EventHandler set
* it is manually updated

This creates a very generic RawEventListener API for receiving events. Any
number of those can then be added to the list for receiving events. Then
this also generates a higher level EventListener abstract class which breaks
out all the event types into abstract methods to implement.  The method
names are auto-generated based on the names in the Tor source code, so the
camelCasing is sometimes weird.